refactor(limit): extract the abuse budget into src/limit.py
The token-bucket core (take/refund), client_ip and its trusted-header/
proxy-evidence logic, the 429/announcement renderers (refill_rate,
limited, budget_note), _settle_room_budget, the long-poll waiter slots
and their state move out of app.py, rationale comments with them.

Knobs arrive as parameters, not module reads: app keeps the module-level
aliases the tests monkeypatch (RATE_*, CLIENT_IP_HEADER, MAX_BUCKETS,
MAX_WAITERS_*) and config.override() mirrors into, and passes them into
limit.take(...)/client_ip(...) exactly as routes already passed
per_min/burst. Mutable state is re-exported as the SAME objects, so
app_module._buckets.clear() and the /stats reader keep working; the
int _waiters_total (rebound by limit's global) resolves through a module
__getattr__. HeaderLimits stays in app: it bounds the header block at
the HTTP edge beside MAX_BODY, not the abuse budget.
